Fix integration tests not to rely on empty stderr
Since now commands executed from within mbt write their error output to
stderr, the stderr might not be empty even when the process is completed
successfully (e.g. npm writes its warnings to srderr). So instead we check
if the command existed successfully and print the stderr output in case
it should have but didn't.